+++ src/components/Footer/Footer.css
@@ -15,18 +15,15 @@
   padding: 0 0 20px;
 }
 
-.footer__project-links {
+.footer__project-info {
+  display: flex;
+  justify-content: space-between;
   padding: 20px 0;
 }
 
 .footer__text {
   margin: 0;
   padding: 0;
-}
-
-.footer__project-links {
-  display: flex;
-  justify-content: space-between;
 }
 
 .footer__links {
@@ -42,9 +39,10 @@
     line-height: calc(15 / 12);
   }
 
-  .footer__project-links {
+  .footer__project-info {
     flex-direction: column-reverse;
     align-items: center;
+    padding: 30px 0 10px;
   }
 
   .footer__links {
@@ -53,7 +51,7 @@
   }
 
   .footer__text {
-    margin: 40px 0 0;
+    margin: 30px 0 0;
     color: #8b8b8b !important;
   }
 }
