+++ src/components/Header/Header.js
@@ -1,23 +1,27 @@
 import React from "react";
+import { useState } from "react";
 import "./Header.css";
-// import { Route, Routes } from "react-router-dom";
-// import HeaderLogIn from "../HeaderLogIn/HeaderLogIn";
+import HeaderLogIn from "../HeaderLogIn/HeaderLogIn";
 import HeaderLogOff from "../HeaderLogOff/HeaderLogOff";
+import Navigation from "../Navigation/Navigation";
 
-export default function Header() {
+export default function Header( { isLoggedIn }) {
+  const [isNavigationOpen, setIsNavigationOpen] = useState(false);
+
+  //Открытие попапов
+  function handleNavigationClick() {
+    setIsNavigationOpen(true);
+  }
+
+  //Закрытие попапов
+  function closeAllPopups() {
+    setIsNavigationOpen(false);
+  }
+
   return (
     <header className="header">
-      <HeaderLogOff />
-
-      {/* Заготовка для JS */}
-      {/* <Routes>
-        <Route exact path="/" element={<HeaderLogIn />} />
-        <Route
-          exact
-          path="/(movies|saved-movies|profile)"
-          element={<HeaderLogIn />}
-        />
-      </Routes> */}
+      {isLoggedIn ?  <HeaderLogIn onNavigation={handleNavigationClick} /> : <HeaderLogOff />}
+      <Navigation isOpen={isNavigationOpen} onClose={closeAllPopups} />
     </header>
   );
 }
