+++ src/components/Main/Main.js
@@ -8,10 +8,10 @@ import AboutMe from "../AboutMe/AboutMe";
 import Portfolio from "../Portfolio/Portfolio";
 import Footer from "../Footer/Footer";
 
-export default function Main() {
+export default function Main({ isLoggedIn }) {
   return (
     <main className="main">
-      <Header />
+      <Header isLoggedIn={isLoggedIn} />
       <Promo />
       <AboutProject />
       <Techs />
