+++ src/components/Navigation/Navigation.js
@@ -1,7 +1,6 @@
 import React from "react";
 import { NavLink } from "react-router-dom";
 import "./Navigation.css";
-import profile from "../../images/profile.svg";
 
 export default function Navigation({ isOpen, onClose }) {
   return (
@@ -11,8 +10,7 @@ export default function Navigation({ isOpen, onClose }) {
           className="button navigation__close-btn"
           type="button"
           onClick={onClose}
-        >
-        </button>
+        ></button>
 
         <nav className="navigation_main-links">
           <NavLink to="/" className="link navigation_link">
@@ -31,13 +29,10 @@ export default function Navigation({ isOpen, onClose }) {
             Аккаунт
           </NavLink>
           <NavLink to="/profile" className="navigation__profile-link">
-            <button className="button navigation__profile-btn">
-              <img
-                className="navigation__logo-profile-btn"
-                src={profile}
-                alt="Лого профиля"
-              ></img>
-            </button>
+            <button
+              className="button navigation__profile-btn"
+              type="button"
+            ></button>
           </NavLink>
         </div>
       </div>
